Move prompt to render module
