+++ examples/ignore_continue/ignore_continue.py
@@ -2,7 +2,7 @@
 
 from simpleline import INPUT_PROCESSED
 from simpleline.base import App, UIScreen
-from simpleline.prompt import Prompt
+from simpleline.render.prompt import Prompt
 from simpleline.widgets import TextWidget, CenterWidget
