+++ simpleline/adv_widgets.py
@@ -17,11 +17,12 @@
 # Red Hat, Inc.
 #
 import sys
+
 from simpleline import INPUT_PROCESSED
 from simpleline import widgets
-from simpleline.prompt import Prompt
-from simpleline.utils.i18n import _, N_, C_
 from simpleline.base import UIScreen
+from simpleline.render.prompt import Prompt
+from simpleline.utils.i18n import _, N_, C_
 
 
 class ErrorDialog(UIScreen):
