+++ simpleline/render/renderer.py
@@ -19,17 +19,17 @@
 # Author(s): Jiri Konecny <jkonecny@redhat.com>
 #
 
-import sys
-import threading
 import getpass
 import queue
-from simpleline.screen_stack import ScreenStack, ScreenData, ScreenStackEmptyException
+import sys
+import threading
+
 from simpleline.event_loop import ExitMainLoop, ExitAllMainLoops
 from simpleline.event_loop.signals import ExceptionSignal, InputReadySignal
-from simpleline.prompt import Prompt
-from simpleline.widgets import TextWidget
 from simpleline.render import RendererUnexpectedError
-
+from simpleline.render.prompt import Prompt
+from simpleline.screen_stack import ScreenStack, ScreenData, ScreenStackEmptyException
+from simpleline.widgets import TextWidget
 
 RAW_INPUT_LOCK = threading.Lock()
